+++ lib/cloud_interactor/domain/update_record.rb
@@ -26,17 +26,31 @@ class CloudInteractor
 
         case @options['route_dns_changes_via']
         when /rackspace|dnsimple/
-          specific_record.type  = args['type']
-          specific_record.value = args['target_ip']
-          specific_record.ttl   = args['ttl']
+          begin
+            specific_record.type  = args['type']
+            specific_record.value = args['target_ip']
+            specific_record.ttl   = args['ttl']
 
-          specific_record.save
+            specific_record.save
+          rescue Fog::Errors::Error => e
+            tries ||= 3
+            puts "Issue updating records for the domain #{ args[IDENTITY.singularize] }! Error: #{e}. Trying #{tries} more times."
+            tries -= 1
+            if tries > 0
+              sleep(15)
+              retry
+            else
+              false
+            end
+          end
         else
           raise "Unsupported action #{ __method__ } for #{ @options['preferred_cloud'] }. Please create an issue on github or submit a PR to fix this issue."
         end
 
         puts "(#{ IDENTITY.capitalize }) Updated #{ args['subdomain'] } (#{ args['target_ip'] }) to #{ args[IDENTITY.singularize] } (#{ args['target_domain'] })..."
       else
+        tries = 3
+        
         create_record [ args ]
       end
     end