+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 2.13.2
+
+* Minor Fix for Rackspace clients that have issues fetching domain record data from Rackspace's domain API
+
+* Added hooks for /var/log/syslog tailing and logging for meteor distributions
+
+* Can now run meteor commands with `cft run` or `cft meteor`
+
+* Added keepalives to all major ssh commands (and set the option to sshkit)
+
+* Added new command `cft backups status` to check the current state of backups being loaded
+
 ## 2.13.1
 
 * Added several new aliases:
@@ -10,7 +22,7 @@
 
     * `cft cb` -> `cft cloud_bootstrap`
 
-    * `cft cl` -> `cft client_lists`
+    * `cft cl` -> `cft client_list`
 
     * `cft e` -> `cft environment`